Improve PostService functionality and update dependencies
Enhanced the functionality of the PostService by adding a method to get posts commented by a user. Refactored the PostType from an export type to constant and updated the usage across the files. The code for the routes has been slightly reformatted. Also, the package-lock.json file was updated to reflect the upgraded versions of the 'express' and 'cookie' packages.

Took 58 minutes